Silence internal libpng build, rename files to stop claiming this is ONScripter

+++ PonscripterLabel_sound.cpp
@@ -21,7 +21,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "ONScripterLabel.h"
+#include "PonscripterLabel.h"
 #if defined(LINUX)
 #include <signal.h>
 #endif
@@ -567,7 +567,7 @@ static size_t oc_read_func(void *ptr, size_t size, size_t nmemb, void *datasourc
     OVInfo *ogg_vorbis_info = (OVInfo*)datasource;
 
     size_t len = size*nmemb;
-    if (ogg_vorbis_info->pos+len > ogg_vorbis_info->length)
+    if (ogg_vorbis_info->pos + len > size_t(ogg_vorbis_info->length))
         len = ogg_vorbis_info->length - ogg_vorbis_info->pos;
     memcpy(ptr, ogg_vorbis_info->buf+ogg_vorbis_info->pos, len);
     ogg_vorbis_info->pos += len;
